style create new competition page and add competition icon
