+++ client/src/App.jsx
@@ -3,7 +3,7 @@ import Layout from "./components/Layout";
 import RequireAuth from "./pages/auth/RequireAuth";
 import Main from "./components/Main";
 import Authorization from "./pages/auth/Authorization";
-import NewCompetition from "./pages/newcompetition/NewCompetition";
+import CreateCompetition from "./pages/newcompetition/CreateCompetition";
 import Profile from "./pages/account/Profile";
 import TournamentPage from "./pages/tournaments/TournamentPage";
 import "./App.css";
@@ -21,7 +21,7 @@ function App() {
           <Route element={<RequireAuth />}>
             <Route path="profile" element={<Profile />} />
             <Route path="tournament/:id" element={<TournamentPage />} />
-            <Route path="new-competition" element={<NewCompetition />} />
+            <Route path="new-competition" element={<CreateCompetition />} />
           </Route>
         </Route>
       </Routes>
