--- client/src/pages/newcompetition/NewCompetition.jsx
@@ -1,13 +1,0 @@
-import Create from "./Create";
-
-function NewCompetition() {
-  return (
-    <div>
-      <p>New Competion Page</p>
-      <h1>Create</h1>
-      <Create />
-    </div>
-  );
-}
-
-export default NewCompetition;
